+++ test/spec/mini_reel.service.ut.js
@@ -403,12 +403,6 @@ define(['app', 'minireel', 'c6uilib', 'angular'], function(appModule, minireelMo
                         expect(result).toBe(angular.copy.calls.mostRecent().result);
                     });
 
-                    it('should give each video a player', function() {
-                        result.forEach(function(video) {
-                            expect(video.player).toBeNull();
-                        });
-                    });
-
                     describe('social', function() {
                         it('should create an array of social links from the links hash', function() {
                             expect(result[0].social).toEqual([]);